[ark-physics-kit] chore: Move ArkSKPhysicsBodyFactory to separate file
